Modularise content.js
