+++ src/content.js
@@ -1,3 +1,5 @@
+import handleScrapedData from "./utilityFunctions/handleScrapedData";
+
 // Placeholder for content script functionality
 console.log("Content script loaded.");
 
@@ -28,59 +30,7 @@ chrome.runtime.sendMessage(
   }
 );
 
-// Function to handle the scraped data - alert user of bad ingredients
-function handleScrapedData(data) {
-  const scrapedItemName = data.itemName || "Unknown Item";
-  const scrapedIngredients = data.ingredientsList || [];
-
-  console.log("Scraped Item Name:", scrapedItemName);
-  console.log("Scraped Ingredients:", scrapedIngredients);
-
-  // Retrieve the bad ingredients from Chrome storage
-  chrome.storage.sync.get("badIngredients", function (result) {
-    let existingIngredients = result.badIngredients || [];
-    let matchedBadIngredients = new Set();
-
-    console.log("Existing Bad Ingredients:", existingIngredients);
-
-    // Create a single regex pattern for all bad ingredients
-    const badIngredientsPattern = existingIngredients
-      .map(escapeRegExp)
-      .join("|");
-    const regex = new RegExp(`\\b(${badIngredientsPattern})\\b`, "gi");
-
-    // Check all ingredients at once
-    scrapedIngredients.forEach((ingredient) => {
-      const matches = ingredient.match(regex);
-      if (matches) {
-        matches.forEach((match) =>
-          matchedBadIngredients.add(match.toLowerCase())
-        );
-      }
-    });
-
-    console.log("Matched Bad Ingredients:", [...matchedBadIngredients]);
-
-    displayResults(scrapedItemName, matchedBadIngredients);
-  });
-}
-
 // Function to escape special characters in regex
 function escapeRegExp(string) {
   return string.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-// Function to display results to the user
-function displayResults(itemName, matchedIngredients) {
-  const badIngredientsArray = [...matchedIngredients];
-
-  if (badIngredientsArray.length > 0) {
-    const alertMessage = `Warning: The product "${itemName}" contains the following concerning ingredients:\n\n${badIngredientsArray.join(
-      ", "
-    )}`;
-    console.log("Displaying dialog with message:", alertMessage);
-    createDialog(alertMessage);
-  } else {
-    console.log(`No concerning ingredients found in "${itemName}".`);
-  }
 }
